backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
python-dotenv==1.0.1
pyodbc==5.1.0
sqlalchemy
pydantic-settings
python-docx
python-multipart